refactor: replace AtomColorLegend + StructureLegend with AtomLegend
- Updated atom color handling logic to streamline the rendering of atom properties.
- fix Wyckoff position based site coloring only taking into account Wyckoff letter, not coloring by Wyckoff orbit
- Adjusted tests to reflect the changes in component names and functionality.
