+++ src/lib/structure/index.ts
@@ -10,17 +10,16 @@ export { default as Bond } from './Bond.svelte'
 export * as bonding_strategies from './bonding'
 
 export { default as Arrow } from './Arrow.svelte'
-export { default as AtomColorLegend } from './AtomColorLegend.svelte'
+export * from './atom-properties'
+export { default as AtomLegend } from './AtomLegend.svelte'
 export { default as CanvasTooltip } from './CanvasTooltip.svelte'
 export { default as Cylinder } from './Cylinder.svelte'
 export { default as Lattice } from './Lattice.svelte'
-export * from './atom-properties'
 export * from './pbc'
 export { default as Structure } from './Structure.svelte'
 export { default as StructureControls } from './StructureControls.svelte'
 export { default as StructureExportPane } from './StructureExportPane.svelte'
 export { default as StructureInfoPane } from './StructureInfoPane.svelte'
-export { default as StructureLegend } from './StructureLegend.svelte'
 export { default as StructureScene } from './StructureScene.svelte'
 export * from './supercell'
